Fix setNumOutputLayerSets() to accept value 256
Max number of output layer sets is 256, which doesn't fit uint8_t.
Change parameter type to uint32_t.

+++ source/Lib/CommonLib/Slice.h
@@ -1093,7 +1093,7 @@ public:
   void              setOlsModeIdc(uint32_t t) { m_vpsOlsModeIdc = t; }
 
   uint32_t          getNumOutputLayerSets() const { return m_vpsNumOutputLayerSets; }
-  void              setNumOutputLayerSets(uint8_t t) { m_vpsNumOutputLayerSets = t; }
+  void              setNumOutputLayerSets(uint32_t t) { m_vpsNumOutputLayerSets = t; }
 
   bool              getOlsOutputLayerFlag(uint32_t ols, uint32_t layer) const { return m_vpsOlsOutputLayerFlag[ols][layer]; }
   void              setOlsOutputLayerFlag(uint32_t ols, uint32_t layer, bool t) { m_vpsOlsOutputLayerFlag[ols][layer] = t; }
